Add files via upload
So for the test file you need to have python and serial intalled with "pip install serial" command and restarting vscode or executing again teh program. dont rename it to serial the tester file cause it wont work and this program was written for a 4 relay arduino shield keep in mind that using any other shield way ground out a pin with the metal usb female port in the arduino giving you an always open setting. For wiring use the NC (Normally closed) and Com pin for anything you want to control.